Renames GithubAccess to GithubBot
`GithubBot` is symantically the better name and improves test isolation. Before the client used by GithubAccess would be a connected client which requires are real private key and app id.
With the client being being a constructor parameter for the GithubBot it is now possible to mock the client and not rely on real credentials for the bot.