Fixes `Build` `tag_name` logic
Previously it was possible that the tag name would be generated as `1.2.3-build.123-build.124`. We now check if the version already contains a build component and only append one if it doesn't exist.
On a release branch this would never be a problem as the use the format `release/1.2.3` and therefore the version is taken as `1.2.3`. But if at a later point we would like to use the same `tag_name` method from develop, this is issue would start appearing.